Add three functions to improve smartparens experience with org-mode.
+org-sp-in-latex-braces-p detects when the cursor is inside LaTeX
curly braces, such as \begin{ ... }.
This is added for the asterisk (*) so you can type something like
\begin{equation*} without having an extra asterisk.

+org-sp-in-setting-decl-p detects when you're typing an org settings
declaration, such as #+BEGIN_SRC
This is added for the underscore (_) so you can type #+BEGIN_SRC
without getting an extra underscore.

+org-sp-in-latex-block-p detects when you're inside a LaTeX block,
such as an equation (\begin{equation} ... \end{equation}).
This is added for the slash (/) to prevent you from getting an
extra / when you are trying to type ratios.